chore: update web components (#568) (#582)
* chore: use Pyodide `0.23.1`

* docs: copy `design_space_exploration` notebook to jupyter lite

* chore: update project classifiers

* docs: update `design_space_exploration` notebook in jupyter lite

* docs: add favicon to jupyter book site

* docs: update `draco-pyodide` npm readme with examples

* feat: auto-generate Jupyter Lite contents from local docs

* fix: install `toml` and `ipywidgets` in debug notebook on demand

This is needed for the Jupyter Lite app to work properly. Instead of
bloating the pyodide distro with `toml` and `ipywidgets` just for
one notebook, we just lazily install these inside the notebook.

* fix: include data assets into the Jupyter Lite `files` dir

* refactor: use `importlib.resources` to load `.lp` files in notebooks

This way, we rely on the already installed `draco` package, instead
of our local file system. This change was needed for Jupyter Lite to
access example `.lp` files without having to host the whole
`draco` package source.